requirements.txt
Django==3.0.5
djangorestframework==3.11.0
python-decouple==3.3
celery[redis]==4.4.2